launchpad/pyproject.toml
[project]
name = "launchpad"
version = "0.1.1"
description = "GPU training orchestrator for Hydrocast"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "httpx>=0.27",
    "pydantic>=2",
    "pyyaml>=6",
    "huggingface-hub>=0.20",
    "python-dotenv>=1.0",
]


[project.scripts]
launchpad = "launchpad.cli:app"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

# -----------------------
# Bump My Version
# -----------------------
[tool.bumpversion]
current_version = "0.1.1"
commit = false
tag = false
allow_dirty = true

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/launchpad/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
